Update core struct (#111)
* Update core struct

* Externalize helpers
