+++ src/components/Router.tsx
@@ -2,8 +2,8 @@ import debug from "@wbe/debug";
 import { BrowserHistory, HashHistory, MemoryHistory } from "history";
 import { Match } from "path-to-regexp";
 import React from "react";
-import { formatRoutes } from "../core/helpers";
-import { getNotFoundRoute, getRouteFromUrl } from "../core/matcher";
+import { formatRoutes } from "../core/core";
+import { getNotFoundRoute, getRouteFromUrl } from "../core/core";
 import { Routers } from "../core/Routers";
 import LangService from "../core/LangService";
 import { staticPropsCache } from "../core/staticPropsCache";
